Rename gammapy/cube/new.py to gammapy/cube/make.py
